third_party/WebKit/LayoutTests/media/track/text-track-cue-exceptions-expected.txt

Verify that TextTrackCue exceptions are properly messaged to developers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cue.startTime = Number.NaN; threw exception TypeError: Failed to set the 'startTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS cue.startTime = Number.POSITIVE_INFINITY; threw exception TypeError: Failed to set the 'startTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS cue.startTime = Number.NEGATIVE_INFINITY; threw exception TypeError: Failed to set the 'startTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS cue.endTime = Number.NaN; threw exception TypeError: Failed to set the 'endTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS cue.endTime = Number.POSITIVE_INFINITY; threw exception TypeError: Failed to set the 'endTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS cue.endTime = Number.NEGATIVE_INFINITY; threw exception TypeError: Failed to set the 'endTime' property on 'TextTrackCue': The provided double value is non-finite..
PASS successfullyParsed is true

TEST COMPLETE
